Add files via upload
The Global Database on Intergenerational Mobility (GDIM) contains estimates of intergenerational mobility (IGM) in education by 10-year cohorts, covering individuals born between 1940 and 1989. IGM is the extent to which living standards of a generation are higher than those of their parents or the extent to which an individual’s position on the socio-economic scale is independent of the position of his or her parents.